Introduce aggressive auto-locking and energy saving
This uses `acpilight` instead of `xbacklight` because that works more
systems with backlight, not only Intel™®© ones. Thus, udev rules which
make the files writeable for users in the `video` group are included.
Logind's idle mechanism only works correctly on X11 sessions. Because I
currently start the UI using good ol' `startx` this misbehaves and
thinks the system is always idle. A getty override forces the session
type to be always X11.
Finally, the deprecated delayed-hibernation service from
https://github.com/Chrysostomus/delayed-hibernation
is included because systemd's `suspend-then-hibernate` is quite funky
for me right now and doesn't really work on all machines.