Introduce wallpaper locking
Also globalize wallpaper script (can still be overwritten).

If an argument ends with `!`, the wallpaper is locked. This means that
if the wallpaper script is called without an argument (not an empty
argument) as the 'wallpaper' and by extension 'output' device commands
do, doesn't change the wallpapers. This means that changing the outputs
doesn't randomize the outputs (the device specific wallpaper script did
that before). If the argument is only `!`, the current wallpaper is
locked (without changing it). In the case one still wants to shuffle the
wallpapers before locking, `?!` can be used. An empty argument (or one
not ending with `!`) resets any locking and selects new wallpapers
according to the argument (and empty argument selects randomly from the
complete pool, except files starting with `00`).