Add over-engined backlight script
Supports external displays using DDC/CI, thus the ability is available
on all devices, not only ones with backlight. All devices are controlled
simultaneously, which also improves consistency.
This uses the `ddcci-driver-linux-dkms` arch package and assumes
`acpilight` is used as an `xbacklight` backend. A new service to set up
some data for the user processes. The DDC/CI interface is rather slow on
my devices so the time is adjusted on the fly (using a cache built by
the service) to ensure that the actual execution time stays roughly the
same.
One can request updates service using a named pipe. Could I have been
using sockets for this? Who knows!
Also add new keybindings to i3 to take advantage of the new
universality.

* Fix xidlelock script not using the new file paths (after the great
  denogrimification)