Add running video detection and prevent screen darkening
For a while now videos in my browser don't block the screensaver
anymore… I wanted a method to detect when a video is playing and
preventing screen darkening (because relying on the 'something is
fullscreened'-heuristic is annoying in i3/sway) from the moment I
introduced the aggressive saving techniques so now there is an overly
complex tridactyl script to take care of this…
Unlike the native browser screensaver block this script allows for more
gradual control. This means that buffering now “re-enables” the
screensaver such that endless buffering doesn't stall the screensaver.
If the tab is unfocussed or the video is audio-only the screensaver will
also be enabled. (The audio-only detection is currently hard-coded based
on specific cases).
Other players on my machine (i.e. `mpv`, `mplayer`) seem to already
block the screensaver by resetting the X idle timer so no need to change
anything there (for now)

* Fix brightness.sh reporting the wrong number of arguments used