Change screen saver handler again
As the old implementation didn't work with the CSP  of some sites.
Tampermonkey's `GM_xmlhttpRequest` is not blocked but gets cancelled
when the tab is closing. `navigator.sendBeacon` is also cancelled and
blocked by uBlock and CSP so no option here… Now both `tridactyl` and
Tampermonkey join forces and the “““REST””” server is put out of a job
because now I use the power of the `postMessage` API to contact
`tridactyl` and let it do what the first version did.
Yay.
Will this be the last time I try to fix this? Who knows!
Please revive me.